Improvements to Array extension and naming.
