+++ AirCasting/Extensions/Numeric+Decrement.swift
@@ -1,0 +1,10 @@
+// Created by Lunar on 27/09/2022.
+//
+
+import Foundation
+
+extension Numeric {
+    func decrement() -> Self {
+        self - 1
+    }
+}
